+++ frontend/src/components/configEditor.js
@@ -40,6 +40,7 @@ import { Templates } from '@rjsf/mui';
 // Frontend custom components
 import { getCssVariable } from './muiTheme';
 import { selectDevices, selectEntities } from './Home';
+import { debug } from '../App';
 
 const { BaseInputTemplate } = Templates; 
 
@@ -216,10 +217,11 @@ export function createConfigTheme(primaryColor) {
 
 export function ArrayFieldTemplate(props) {
   const { canAdd, onAddClick, schema, title } = props;
-  // console.log('ArrayFieldTemplate: title', title, 'description', schema.description, 'items', props.items);
+  if(debug) console.log('ArrayFieldTemplate: title', title, 'description', schema.description, 'items', props.items);
 
   const [dialogDeviceOpen, setDialogDeviceOpen] = useState(false);
   const [dialogEntityOpen, setDialogEntityOpen] = useState(false);
+  const [dialogDeviceEntityOpen, setDialogDeviceEntityOpen] = useState(false);
   const [filter, setFilter] = useState('');
 
   const primaryColor = useMemo(() => getCssVariable('--primary-color', '#009a00'), []);
@@ -235,6 +237,10 @@ export function ArrayFieldTemplate(props) {
 
   const handleDialogEntityToggle = () => {
     setDialogEntityOpen(!dialogEntityOpen);
+  };
+
+  const handleDialogDeviceEntityToggle = () => {
+    setDialogDeviceEntityOpen(!dialogDeviceEntityOpen);
   };
 
   const handleSelectDeviceValue = (value) => {
@@ -259,6 +265,17 @@ export function ArrayFieldTemplate(props) {
     onAddClick();
   }
 
+  const handleSelectDeviceEntityValue = (value) => {
+    // console.log('ArrayFieldTemplate: handleSelectEntityValue', value);
+    setDialogDeviceEntityOpen(false);
+    // Trigger onAddClick to add the selected new item
+    if(schema.selectDeviceEntityFrom === 'name')
+      schema.items.default = value.name;
+    else if(schema.selectDeviceEntityFrom === 'description')
+      schema.items.default = value.description;
+    onAddClick();
+  }
+
   return (
     <Box sx={{ padding: '10px', margin: '0px', border: '1px solid grey' }}>
       {title && (
@@ -278,6 +295,13 @@ export function ArrayFieldTemplate(props) {
               {schema.selectEntityFrom && 
                 <Tooltip title="Add an entity from the list">
                   <IconButton onClick={handleDialogEntityToggle}>
+                    <ListIcon />
+                  </IconButton>
+                </Tooltip>
+              }
+              {schema.selectDeviceEntityFrom && 
+                <Tooltip title="Add a device entity from the list">
+                  <IconButton onClick={handleDialogDeviceEntityToggle}>
                     <ListIcon />
                   </IconButton>
                 </Tooltip>
@@ -387,6 +411,36 @@ export function ArrayFieldTemplate(props) {
             <Button onClick={handleDialogEntityToggle}>Close</Button>
           </DialogActions>
         </Dialog>
+
+        {/* Dialog for selecting a device entity */}
+        <Dialog open={dialogDeviceEntityOpen} onClose={handleDialogDeviceEntityToggle} PaperProps={{
+            sx: {
+              maxHeight: '50vh', // Set the maximum height to 50% of the viewport height
+              overflow: 'auto',  // Allow scrolling for overflowing content
+            },
+          }}>
+          <DialogTitle>Select an entity for {title}</DialogTitle>
+          <DialogContent>
+            <List dense>
+              {selectDevices.filter((d) => d.serial === title).map((value, index) => {
+                console.log('ArrayFieldTemplate: handleSelectDeviceEntityValue value:', value, value.entities);
+                return value.entities.map((entity, index) => (
+                  <ListItemButton onClick={() => handleSelectDeviceEntityValue(entity)} key={index}>
+                    {entity.icon==='wifi' && <ListItemIcon><WifiIcon /></ListItemIcon>}
+                    {entity.icon==='ble' && <ListItemIcon><BluetoothIcon /></ListItemIcon>}
+                    {entity.icon==='hub' && <ListItemIcon><HubIcon /></ListItemIcon>}
+                    {entity.icon==='component' && <ListItemIcon><ViewInArIcon /></ListItemIcon>}
+                    {entity.icon==='matter' && <ListItemIcon><DeviceHubIcon /></ListItemIcon>}
+                    <ListItemText primary={entity.name} secondary={entity.description}/>
+                  </ListItemButton>
+                ));
+              })}
+            </List>
+          </DialogContent>
+          <DialogActions>
+            <Button onClick={handleDialogDeviceEntityToggle}>Close</Button>
+          </DialogActions>
+        </Dialog>
       </ThemeProvider>
 
     </Box>
@@ -394,11 +448,71 @@ export function ArrayFieldTemplate(props) {
 }
 
 export function ObjectFieldTemplate(props) {
-  const { onAddClick, schema, properties, title, description } = props;
+  const { onAddClick, schema, properties, title, description, formData, registry } = props;
+
+  const [dialogDeviceOpen, setDialogDeviceOpen] = useState(false);
+  const [filter, setFilter] = useState('');
+  const [newkey, setNewkey] = useState('');
+
+  const primaryColor = useMemo(() => getCssVariable('--primary-color', '#009a00'), []);
+  const theme = useMemo(() => createConfigTheme(primaryColor), []);
+
+  const handleFilterChange = (event) => {
+    setFilter(event.target.value);
+  };
+
+  const handleDialogDeviceToggle = () => {
+    setDialogDeviceOpen(!dialogDeviceOpen);
+  };
+
+  const handleSelectDeviceValue = (value) => {
+    // console.log('ObjectFieldTemplate: handleSelectValue value:', value);
+    // console.log('ObjectFieldTemplate: handleSelectValue schema before:', schema);
+    setDialogDeviceOpen(false);
+    let newkey = '';
+    if(schema.selectFrom === 'serial')
+      newkey = value.serial;
+    else if(schema.selectFrom === 'name')
+      newkey = value.name;
+    setNewkey(newkey);
+    // console.log('ObjectFieldTemplate: handleSelectValue newkey:', newkey);
+    // schema.properties[newkey] = { description: 'New device', type: 'array', uniqueItems: true, items: { type: 'string' }, selectEntityFrom: "name" };
+    // Trigger onAddClick returned function to add the selected new item
+    const addProperty = onAddClick(schema);
+    addProperty();
+    // console.log('ObjectFieldTemplate: handleSelectValue schema after:', schema);
+    // console.log('ObjectFieldTemplate: handleSelectValue schema after:', properties);
+  };
+
+  const handleAddItem = (event) => {
+    // console.log('ObjectFieldTemplate: handleAddItem schema before:', schema, formData, registry);
+    // console.log('ObjectFieldTemplate: handleSelectValue properties before:', properties);
+    // Trigger onAddClick returned function to add the selected new item
+    const addProperty = onAddClick(schema);
+    addProperty();
+    // console.log('ObjectFieldTemplate: handleAddItem schema after:', schema, formData, registry);
+    // console.log('ObjectFieldTemplate: handleSelectValue properties after:', properties);
+  };
+
   // Check if this is the entire schema or an individual object
   const isRoot = !schema.additionalProperties;
-  console.log('ObjectFieldTemplate: title', title, 'description', description, 'schema', schema, 'isRoot', isRoot);
+  if(debug) console.log('ObjectFieldTemplate: title', title, 'description', description, 'schema', schema, 'isRoot', isRoot);
+  if(debug) console.log('ObjectFieldTemplate: props', props);
+
+  if(!isRoot) console.log('ObjectFieldTemplate: properties', properties);
+  /*
   console.log('ObjectFieldTemplate: properties', properties);
+  properties.forEach((p) => {
+    console.log('ObjectFieldTemplate: ', p.name);
+    if(p.name==='newKey' && newkey!=='') {
+      console.log('ObjectFieldTemplate: changed', p.name, 'to', newkey);
+      p.name = newkey;
+      p.content.key = newkey;
+      p.content.props.name = newkey;
+      setNewkey('');
+    }
+  });
+  */
 
   return (
     <Box sx={{ padding: '10px', margin: '0px', border: isRoot ? 'none' : '1px solid grey' }}>
@@ -412,11 +526,20 @@ export function ObjectFieldTemplate(props) {
       {title && !isRoot && (
         <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', padding: '0px', margin: '0px', marginBottom: '0px' }}>
           <Typography sx={titleSx}>{title}</Typography>
-          <Tooltip title="Add a new item">
-            <IconButton onClick={onAddClick(schema)} size="small" color="primary">
-              <Add />
-            </IconButton>
-          </Tooltip>
+          <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', padding: '0px', margin: '0px', marginBottom: '0px' }}>
+            {schema.selectFrom && 
+              <Tooltip title="Add a device from the list">
+                <IconButton disabled onClick={handleDialogDeviceToggle}>
+                  <ListIcon />
+                </IconButton>
+              </Tooltip>
+            }
+            <Tooltip title="Add a new item">
+              <IconButton onClick={handleAddItem} size="small" color="primary">
+                <Add />
+              </IconButton>
+            </Tooltip>
+          </Box>
         </Box>
       )}
       {/* Description for both */}
@@ -445,6 +568,44 @@ export function ObjectFieldTemplate(props) {
           </Box>
         </Box>
       ))}
+
+      <ThemeProvider theme={theme}>
+        {/* Dialog for selecting a device */}
+        <Dialog open={dialogDeviceOpen} onClose={handleDialogDeviceToggle} PaperProps={{
+            sx: {
+              maxHeight: '50vh', // Set the maximum height to 50% of the viewport height
+              overflow: 'auto',  // Allow scrolling for overflowing content
+            },
+          }}>
+          <DialogTitle>Select a device</DialogTitle>
+          <DialogContent>
+            <Box sx={{ display: 'flex', flexDirection: 'row', alignItems: 'center', gap: '10px', marginBottom: '10px' }}>
+              <Typography variant="subtitle1" sx={{ whiteSpace: 'nowrap' }}>Filter by:</Typography>
+              <TextField
+                fullWidth
+                variant="outlined"
+                value={filter}
+                onChange={handleFilterChange}
+                placeholder="Enter serial or name"
+              />
+            </Box>
+            <List dense>
+              {selectDevices.filter((v) => v.serial.includes(filter) || v.name.includes(filter)).map((value, index) => (
+                <ListItemButton onClick={() => handleSelectDeviceValue(value)} key={index}>
+                  {value.icon==='wifi' && <ListItemIcon><WifiIcon /></ListItemIcon>}
+                  {value.icon==='ble' && <ListItemIcon><BluetoothIcon /></ListItemIcon>}
+                  {value.icon==='hub' && <ListItemIcon><HubIcon /></ListItemIcon>}
+                  <ListItemText primary={value.serial} secondary={value.name}/>
+                </ListItemButton>
+              ))}
+            </List>
+          </DialogContent>
+          <DialogActions>
+            <Button onClick={handleDialogDeviceToggle}>Close</Button>
+          </DialogActions>
+        </Dialog>
+      </ThemeProvider>
+
     </Box>
   );
 }