+++ src/matterbridgePlatform.ts
@@ -65,7 +65,7 @@ export class MatterbridgePlatform {
   public version = ''; // Will be set by the loadPlugin() method using the package.json value.
   public storage: NodeStorageManager | undefined;
   public context: NodeStorage | undefined;
-  public selectDevice = new Map<string, { serial: string; name: string; icon?: string; entities?: string[] }>();
+  public selectDevice = new Map<string, { serial: string; name: string; icon?: string; entities?: { name: string; description: string; icon?: string }[] }>();
   public selectEntity = new Map<string, { name: string; description: string; icon?: string }>();
 
   /**