+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Matterbridge edge is now released. The default mode is still the normal mode to 
 The frontend has a new dark and light mode. The dark mode is now the default mode.
 It is possible to change the mode (Classic, Dark or Light) in Settings, Matterbridge settings.
 
-## [1.7.2-dev.6] - 2025-01-10
+## [1.7.2-dev.7] - 2025-01-10
 
 ### Added
 
@@ -29,6 +29,8 @@ It is possible to change the mode (Classic, Dark or Light) in Settings, Matterbr
 - [frontend]: Added the possibility to reorder the items in the config editor lists.
 - [frontend]: Added custom error messages for ErrorListTemplate and FieldErrorTemplate in react-jsonschema-form for validation in the config editor.
 - [frontend]: Added filter by device name and serial number to Devices page.
+- [frontend]: Added Icon view to the Devices page (beta).
+- [frontend]: Added the possibility to select the entities/components from a list in the config editor.
 
 ### Changed